refactor: 메인페이지 events 컴포넌트 분리
